Made changes regarding rotation and level creation
The marble now rotates along with the world. This was done because it
was to hard to rotate the level quickly and not have it clip through the
world. To me this seems like a permanent fix as the ball no longer clips
through the level.

Marble now updates its initial starting position when the level is
created using an image.

Changed the platforms and walls from arrays to lists (since we don't
know how people are going to construct their levels). Removed the
"default test" level (still in comments) and made classes that update
the "platform", "walls" and "forces" variables.

Player is now given a top-down perspective on their marble and the
camera stays just above the marble's position.

Added the LevelScanner class along with two test levels, "testlevel.png"
and "testlevelhard.png" (it's really not that hard). This class reads
each pixel in .png files and creates a level out of them. Black pixels
are walls, white pixels are platforms. Blue and red are recognized, but
at the moment they don't do anything major. Any other color is
unrecognized and doesn't add a platform or wall (so they are essentially
just a hole)