The world can no longer be rotated beyond a certain angle.
The perspective can still get a little bit weird sometimes so I added
controls to they arrow keys as well as shift and control just to rotate
the camera. They were mainly meant for debugging but they can be kept in
there since it doesn't seem to harm anything.

The player's reset method is called when the player's Y coordinate drops
below a certain point. Also fixed the transformations of the
_startingPosition.