Win state is detectable.
Right now all it does is reset the player back to the beginning. Change
was made to the OrientedBoundingBox. Added a field that lets it know if
it is a goal box. If it is, then it sets a goal flag on the Marble when
there is a collision between them. This is relayed to the WorldDirector
and allows it to trigger the win state.