Experimental push blocks added to the level scanner
I might have over complicated it, but red values of 42 are a unique
identifier to the LevelScanner now. If the identifier is detected, then
it takes the green value as the angle and the blue value as magnitude.

I removed the hardcoded experimental push blocks and recreated them
using the LevelScanner.