Added scores
The score starts at 500 for the default testing level, or 10 times the
area of the custom level. It counts down every tick until the player
reaches the end, then it resets and the score is displayed at the bottom
of the screen. If the player falls off of the level then the running
score is not reset.

Fixed the LevelSelectPanel's code to be more intuitive.